FastAPI/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy11